+++ bootstrap/providers.php
@@ -3,6 +3,7 @@
 use App\Providers\AppServiceProvider;
 use App\Providers\AuthServiceProvider;
 use App\Providers\CommunicationServiceProvider;
+use App\Providers\Filament\AdminPanelProvider;
 use App\Providers\PaymentServiceProvider;
 
 return [
@@ -10,4 +11,5 @@ return [
     AuthServiceProvider::class,
     PaymentServiceProvider::class,
     CommunicationServiceProvider::class,
+    AdminPanelProvider::class,
 ];